+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+## 1.63.0 (November 29, 2021)
+
+* resource/spotinst_ocean_gke_launch_spec: added support for `scheduling`
+* resource/spotinst_ocean_aws_launch_spec: added support for `scheduling`
+* resource/spotinst_ocean_ecs_launch_spec: added support for `scheduling`
+
 ## 1.62.0 (November 01, 2021)
 
 ENHANCEMENTS: